refactor: rename parser to syntax
