+++ rust/candid/Cargo.toml
@@ -45,9 +45,9 @@ candid_parser = { path = "../candid_parser" }
 bignum = ["dep:num-bigint", "dep:num-traits"]
 printer = ["dep:pretty"]
 value = ["bignum", "printer"]
-parser = []
+syntax = []
 default = ["serde_bytes", "printer", "bignum"]
-all = ["default", "value", "ic_principal/arbitrary", "parser"]
+all = ["default", "value", "ic_principal/arbitrary", "syntax"]
 
 [[test]]
 name = "types"
